refactor: move TruncatedText to libs/ui
