+++ libs/ui/src/index.ts
@@ -16,6 +16,7 @@ export * from './pure/FiatAmount'
 export * from './pure/TokenSymbol'
 export * from './pure/TokenName'
 export * from './pure/TokenAmount'
+export * from './pure/TruncatedText'
 export * from './pure/SelectDropdown'
 export * from './pure/Tooltip'
 export * from './pure/Popover'
