+++ libs/ui/src/pure/TruncatedText/index.tsx
